requirements.txt
Flask
spacy
nltk
scikit-learn
PyPDF2
pdfplumber
python-docx
docx2txt
textstat
reportlab
python-dateutil
regex